requirements.txt
flask
openai
pdf2image
numpy
opencv-python-headless
lmdeploy
pillow
timm
